+++ functions/PSCalendarConfigurations.ps1
@@ -12,12 +12,13 @@ Function Export-PSCalendarConfiguration {
     Write-Verbose "Using PowerShell version: $($PSVersionTable.PSVersion)"
     Write-Verbose "Running in PowerShell host: $($host.name)"
 
+    #$configPrefPath is a module-scoped variabledir
     Write-Verbose "Exporting configuration to $configPrefPath"
     #$PSCalendarConfiguration and  $configPath are module scoped variables defined in the root module
     $PSCalendarConfiguration | ConvertTo-Json |
     Out-File -FilePath $configPrefPath -Force -Encoding UTF8
     if ($Passthru -AND (-Not $WhatIfPreference)) {
-        Get-Item $ConfigPrefPath
+        Get-Item $ConfigPrefPath -force
     }
     Write-Verbose "Ending: $($MyInvocation.MyCommand)"
 }